+++ core/src/autogluon/core/utils/__init__.py
@@ -1,6 +1,5 @@
 from .files import download, unzip
 from .miscs import *
 from .plots import *
-from .try_import import *
 from .utils import *
 from .version_utils import show_versions
