+++ tabular/src/autogluon/tabular/models/lgb/lgb_model.py
@@ -15,7 +15,7 @@ from autogluon.common.utils.resource_utils import ResourceManager
 from autogluon.core.constants import BINARY, MULTICLASS, REGRESSION, SOFTCLASS
 from autogluon.core.models import AbstractModel
 from autogluon.core.models._utils import get_early_stopping_rounds
-from autogluon.core.utils import try_import_lightgbm
+from autogluon.common.utils.try_import import try_import_lightgbm
 
 from . import lgb_utils
 from .hyperparameters.parameters import get_param_baseline, get_lgb_objective, DEFAULT_NUM_BOOST_ROUND
