+++ tabular/src/autogluon/tabular/models/lgb/lgb_utils.py
@@ -4,7 +4,7 @@ import numpy as np
 from pandas import DataFrame, Series
 
 from autogluon.core.constants import BINARY, MULTICLASS, REGRESSION, SOFTCLASS
-from autogluon.core.utils import try_import_lightgbm
+from autogluon.common.utils.try_import import try_import_lightgbm
 
 
 # Mapping to specialized LightGBM metrics that are much faster than the standard metric computation
