+++ tabular/src/autogluon/tabular/models/lr/lr_rapids_model.py
@@ -3,7 +3,7 @@ import logging
 import numpy as np
 
 from autogluon.core.constants import REGRESSION
-from autogluon.core.utils.try_import import try_import_rapids_cuml
+from autogluon.common.utils.try_import import try_import_rapids_cuml
 
 from .hyperparameters.parameters import get_param_baseline
 from .lr_model import LinearModel
